Architecture updates, and some existing bug fixing with cursors not reverting when button loses focus.

+++ src/main/java/model/Item.java
@@ -14,6 +14,6 @@ public abstract class Item implements Serializable {
     protected double distance(Point point1, Point point2) {
         double xDifference = point1.getX() - point2.getX();
         double yDifference = point1.getY() - point2.getY();
-        return ((double) (Math.sqrt(xDifference * xDifference + yDifference * yDifference)));
+        return Math.sqrt(xDifference * xDifference + yDifference * yDifference);
     }
 }
